+++ sismic/testing/__init__.py
@@ -1,3 +1,0 @@
-from .tester import teststory_from_trace, ExecutionWatcher
-
-__all__ = ['ExecutionWatcher', 'teststory_from_trace']